requirements.txt
prometheus-client==0.14.1
speedtest-cli==2.1.3
